Barebones site without Ethereum interaction
This site has an index page with a button to make an event and a placeholder demonstrating what would be in recent events, a create.html page to enter information for an event, then an event page that show the accepted information and displays a QR code with the info encoded. This is not how the "final POC" will work as the information will be sent through the blockchain then the QR code will only include a URL, but it's a display of the software needed for the final.