Website with Database
QR code on this isn't working, the location/unique message-leaving is not implemented, and the time restrictions for leaving a message are not yet implemented; then of course communication with the blockchain. QR code should be easy I just don't feel like working more tonight. We need to figure out what we're doing about leaving messages and how to implement that. The time restrictions should be fairly easy with a dropdown menu. 

A demonstration of this version with working database and all is up at: forster.tech.